+++ src/utils/get_validator_status.ts
@@ -4,19 +4,28 @@
  * @param jailed boolean
  * @returns an object with status and theme
  */
-export const getValidatorStatus = (status: number, jailed: boolean) => {
+export const getValidatorStatus = (status: number, jailed: boolean, tombstoned: boolean) => {
   const results = {
     status: 'na',
     theme: 'zero',
   };
+  // jailed and tombstone statuses are prioritised over their unbonding state
+  if (jailed) {
+    results.status = 'jailed';
+    results.theme = 'two';
+    return results;
+  }
+
+  if (tombstoned) {
+    results.status = 'tombstoned';
+    results.theme = 'two';
+    return results;
+  }
 
   if (status === 3) {
     results.status = 'active';
     results.theme = 'one';
-  } else if (status === 2 && jailed) {
-    results.status = 'jailed';
-    results.theme = 'two';
-  } else if (status === 2 && !jailed) {
+  } else if (status === 2) {
     results.status = 'unbonding';
     results.theme = 'three';
   } else if (status === 1) {